refactor: 폴더 명 변경
