+++ src/services/lifecycleManager.ts
